apps/api/requirements.txt
# Web Framework
uvicorn[standard]
fastapi
requests
python-multipart 

# Database and Storage
pymongo
python-dotenv
azure-storage-blob

# Image Processing
Pillow
pillow-heif
piexif
exifread



## Misc
simplekml

## Testing
pytest
pytest-mock
